fix(db, stocks): sqlx migrations + Decimal money math
Two intertwined DB infrastructure changes shipped together because
they both restructure src/db/ and the second's Decimal types live
in models the first migrates.

==================================================================
Part 1 (Tier 5.5): sqlx migration system
==================================================================

The hand-rolled CREATE TABLE IF NOT EXISTS sequence in init_pool
had no version tracking, no ALTER path. Adding any column in the
future would either need manual operator intervention or silently
break instances.

Replaced with sqlx::migrate!. The init DDL moves into
migrations/20260414000000_init.sql (every CREATE keeps IF NOT
EXISTS so existing prod databases are no-op'd; sqlx still records
the migration in _sqlx_migrations inside each instance's schema).

init_pool now: opens connection, runs CREATE SCHEMA, builds pool
with after_connect setting search_path, calls
sqlx::migrate!("./migrations").run(&pool). Per-instance
search_path means per-instance _sqlx_migrations table —
migration history is tracked separately per instance, which
matches the schema-per-instance isolation model.

Added 'migrate' to sqlx features.

==================================================================
Part 2 (Tier 5.6): Float -> Decimal for stocks
==================================================================

Stock cash, quantities, average cost, transaction totals were all
f64 / DOUBLE PRECISION. Drift accumulated cents over many trades;
the remaining < 0.0001 dust check was a fragile float-epsilon
guard.

Migrated to rust_decimal::Decimal:

- New migration migrations/20260414000001_stocks_decimal.sql ALTERs
  stock_portfolios.cash_balance, stock_holdings.{quantity,
  avg_cost}, stock_transactions.{quantity, price_per_share,
  total_amount} from DOUBLE PRECISION to NUMERIC(18, 4) via
  USING ::NUMERIC(18, 4). stock_price_cache is left as f64
  (display-only, 60s TTL, converted at the Rust boundary in
  stocks/api.rs).
- Model structs in src/db/models.rs: those fields now Decimal.
- buy_stock / sell_stock in src/db/queries.rs: take/return Decimal,
  dust check is now remaining.is_zero().
- src/commands/stocks.rs: input parsing via Decimal::from_str,
  leaderboard sort via Ord (no more partial_cmp).
- src/stocks/{api, embeds, mod}.rs: all formatters take Decimal;
  format_money / format_pnl / pnl_color all Decimal-typed.
- src/ai/deepseek.rs: only stock-tool dispatch arms touched
  (stock_buy, stock_sell, stock_leaderboard); JSON args
  funnel through Decimal::from_f64 at the AI/Decimal boundary.
- STARTING_CASH is now const Decimal (Decimal::from_parts is
  const fn — no LazyLock needed).

Decimal-vs-f64 caveats: format strings can't use {:.2} with
Decimal, replaced with .round_dp(2) at display boundaries;
divisions get .round_dp(4) to fit NUMERIC(18, 4) precision;
Decimal::ONE_HUNDRED for percentage math.

Added rust_decimal dep with db-postgres and macros features;
added 'rust_decimal' to sqlx features for the type integration.

==================================================================
Verification
==================================================================

cargo fmt clean. cargo clippy --all-targets -- -D warnings clean.
cargo test 37/37 passing.

+++ src/db/models.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use sqlx::FromRow;
 
 #[derive(Debug, FromRow)]
@@ -28,7 +29,7 @@ pub struct GuildSettings {
 pub struct StockPortfolio {
 	pub guild_id: String,
 	pub user_id: String,
-	pub cash_balance: f64,
+	pub cash_balance: Decimal,
 	pub created_at: DateTime<Utc>,
 }
 
@@ -39,8 +40,8 @@ pub struct StockHolding {
 	pub guild_id: String,
 	pub user_id: String,
 	pub symbol: String,
-	pub quantity: f64,
-	pub avg_cost: f64,
+	pub quantity: Decimal,
+	pub avg_cost: Decimal,
 }
 
 #[derive(Debug, FromRow)]
@@ -51,9 +52,9 @@ pub struct StockTransaction {
 	pub user_id: String,
 	pub symbol: String,
 	pub action: String,
-	pub quantity: f64,
-	pub price_per_share: f64,
-	pub total_amount: f64,
+	pub quantity: Decimal,
+	pub price_per_share: Decimal,
+	pub total_amount: Decimal,
 	pub created_at: DateTime<Utc>,
 }
